Refactor Stops
