+++ app/dashboard/stops/[id]/edit/page.tsx
@@ -1,11 +1,11 @@
 'use server';
 
-import { fetchStopById } from "./lib/get-stop-action";
 import EditStopForm from "./ui/edit-stop-form";
+import {getStopById} from "@/app/dashboard/stops/[id]/edit/data/get-stop-by-id";
 
 export default async function Page({ params }: { params: { id: string } }) {
     const id = params.id
-    const stop = await fetchStopById(id)
+    const stop = await getStopById(id)
 
     return (
         <>
