+++ app/dashboard/stops/data/delete-stop.ts
